org - encapsulate PlotlyWrapper, goober styles
